Refactor code and create a specialized script for loading the db
